requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
pillow==10.3.0
numpy==1.24.3
pandas==2.0.3

# Machine Learning & Computer Vision
tensorflow==2.15.0
scikit-learn==1.3.2

# OCR
easyocr==1.7.0
pytesseract==0.3.10

# Vector Database & Search
faiss-cpu==1.7.4
chromadb==0.4.22

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
